ecom

+++ routes/web.php
@@ -11,13 +11,10 @@ use App\Http\Controllers\VolumeController;
 use App\Http\Controllers\ReportController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\AuthController;
+use App\Http\Controllers\EcommerceController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\ItemController;
 use App\Http\Controllers\ReviewController;
-
-Route::get('/', function () {
-    return view('welcome');
-})->name('welcome');
 
 Route::middleware([
     'auth:sanctum',
@@ -27,6 +24,10 @@ Route::middleware([
     Route::get('/dashboard', function () {
         return view('dashboard');
     })->name('dashboard');
+});
+
+Route::controller(EcommerceController::class)->group(function () {
+    Route::get('/', 'getMainPage')->name('welcome');
 });
 
 Route::group(['prefix' => 'admin'], function () {
@@ -68,10 +69,10 @@ Route::group(['prefix' => 'admin'], function () {
 
         Route::controller(BannerSettingController::class)->group(function () {
 
-            Route::get('banners', 'getAll')->name('banner-list')->middleware('permission:banner list');
-            Route::get('banners/{id}', 'read')->name('banner-read')->middleware('permission:banner list');
-            Route::post('banners/{id}', 'store')->name('store-banner')->middleware('permission:add banner');
-            Route::get('banners/delete/{id}', 'delete')->name('delete-banner')->middleware('permission:delete banner');
+            Route::get('banners', 'getAll')->name('banner-list')->middleware(['auth:admin', 'permission:banner list,admin']);
+            Route::get('banners/{id}', 'read')->name('banner-read')->middleware(['auth:admin', 'permission:banner list,admin']);
+            Route::post('banners/{id}', 'store')->name('store-banner')->middleware(['auth:admin', 'permission:add banner,admin']);
+            Route::get('banners/delete/{id}', 'delete')->name('delete-banner')->middleware(['auth:admin', 'permission:delete banner,admin']);
 
         });
 
@@ -83,12 +84,12 @@ Route::group(['prefix' => 'admin'], function () {
 
         Route::controller(AdminController::class)->group(function () {
 
-            Route::get('admins', 'getAll')->name('admin-list')->middleware('permission:user list');
-            Route::get('admins/read/{id}', 'read')->name('read-admin-view')->middleware('permission:user information');
-            Route::get('admins/store', 'createView')->name('create-admin-view')->middleware('permission:add user');
-            Route::post('admins', 'store')->name('create-admin')->middleware('permission:add user');
-            Route::get('admins/change-status/{id}', 'updateStatus')->name('change-status')->middleware('permission:activate/deactivate user');
-            Route::get('admins/delete/{id}', 'delete')->name('admin-delete')->middleware('permission:delete user');
+            Route::get('admins', 'getAll')->name('admin-list')->middleware(['auth:admin', 'permission:user list,admin']);
+            Route::get('admins/read/{id}', 'read')->name('read-admin-view')->middleware(['auth:admin', 'permission:user information,admin']);
+            Route::get('admins/store', 'createView')->name('create-admin-view')->middleware(['auth:admin', 'permission:add user,admin']);
+            Route::post('admins', 'store')->name('create-admin')->middleware(['auth:admin', 'permission:add user,admin']);
+            Route::get('admins/change-status/{id}', 'updateStatus')->name('change-status')->middleware(['auth:admin', 'permission:activate/deactivate user,admin']);
+            Route::get('admins/delete/{id}', 'delete')->name('admin-delete')->middleware(['auth:admin', 'permission:delete user,admin']);
 
         });
 
@@ -103,35 +104,36 @@ Route::group(['prefix' => 'admin'], function () {
         Route::controller(CategoryController::class)->group(function () {
 
             Route::get('category-list', 'index');
-            Route::get('categories', 'getAll')->name('show-categories')->middleware('permission:genre list');
-            Route::put('categories/re-shuffle', 'shuffle')->name('shuffle-categories')->middleware('permission:reshuffle genre');
+            Route::get('categories', 'getAll')->name('show-categories')->middleware(['auth:admin', 'permission:genre list,admin']);
+            Route::put('categories/re-shuffle', 'shuffle')->name('shuffle-categories')->middleware(['auth:admin', 'permission:reshuffle genre,admin']);
         });
 
         Route::controller(ItemController::class)->group(function () {
 
-            Route::get('series', 'getAll')->name('show-items')->middleware('permission:series list');
-            Route::get('series/create', 'createView')->name('create-item-view')->middleware('permission:add series');
-            Route::get('series/read/{id}', 'read')->name('read-item-view')->middleware('permission:add series');
-            Route::post('series', 'create')->name('create-item')->middleware('permission:add series');
-            Route::post('series/{id}', 'update')->name('update-item')->middleware('permission:update series');
-            Route::delete('series/{id}', 'delete')->name('delete-item')->middleware('permission:delete series');
+            Route::get('series', 'getAll')->name('show-items')->middleware(['auth:admin', 'permission:series list,admin']);
+            Route::get('series/create', 'createView')->name('create-item-view')->middleware(['auth:admin', 'permission:add series,admin']);
+            Route::get('series/read/{id}', 'read')->name('read-item-view')->middleware(['auth:admin', 'permission:add series,admin']);
+            Route::post('series', 'create')->name('create-item')->middleware(['auth:admin', 'permission:add series,admin']);
+            Route::post('series/{id}', 'update')->name('update-item')->middleware(['auth:admin', 'permission:update series,admin']);
+            Route::delete('series/{id}', 'delete')->name('delete-item')->middleware(['auth:admin', 'permission:delete series,admin']);
 
         });
 
         Route::controller(VolumeController::class)->group(function () {
 
-            Route::get('volumes', 'getAll')->name('show-volumes');
-            Route::get('volumes/read/{id}', 'read')->name('read-volume-view');
-            Route::get('volumes/change_status/{id}', 'changeStatus');
-            Route::get('volumes/create', 'createView')->name('create-volume-view');
-            Route::post('volumes', 'create')->name('create-volume');
-            Route::put('volumes/update/{id}', 'update')->name('update-volume');
-            Route::get('series/volumes/{id}', 'volumeList');
+            Route::get('volumes', 'getAll')->name('show-volumes')->middleware(['auth:admin', 'permission:volume list,admin']);
+            Route::get('volumes/read/{id}', 'read')->name('read-volume-view')->middleware(['auth:admin', 'permission:volume list,admin']);
+            Route::get('volumes/change_status/{id}', 'changeStatus')->middleware(['auth:admin', 'permission:update volume,admin']);
+            Route::get('volumes/create', 'createView')->name('create-volume-view')->middleware(['auth:admin', 'permission:add volume,admin']);
+            Route::post('volumes', 'create')->name('create-volume')->middleware(['auth:admin', 'permission:add volume,admin']);
+            Route::put('volumes/update/{id}', 'update')->name('update-volume')->middleware(['auth:admin', 'permission:update volume,admin']);
+            Route::get('series/volumes/{id}', 'volumeList')->middleware(['auth:admin', 'permission:delete volume,admin']);
+
         });
 
         Route::controller(ReviewController::class)->group(function () {
 
-            Route::get('review/delete/{id}', 'delete')->name('review-delete')->middleware('permission:delete-review');
+            Route::get('review/delete/{id}', 'delete')->name('review-delete')->middleware(['auth:admin', 'permission:delete-review,admin']);
         });
 
         Route::controller(ReportController::class)->group(function () {
